Partie1/requirements.txt
numpy
opencv-python
scipy
PyQt5
pillow